For some reason the TextFieldWithPickerAsInputView view refuses to be recognized unless it's in the same file as the line calling it. Can't figure out why.
